refactor(right-panel): refactors right panel into separate sections for easy changes
