+++ src/views/wip/sections/__init__.py
@@ -1,0 +1,11 @@
+from ._collapsible import _CollapsibleSection
+from .navigator import DataNavigatorSection
+from .classes import ClassesSection
+from .annotations import AnnotationsSection
+
+__all__ = [
+    "_CollapsibleSection",
+    "DataNavigatorSection",
+    "ClassesSection",
+    "AnnotationsSection",
+]
