+++ src/views/wip/sections/_collapsible.py
@@ -1,0 +1,46 @@
+from PySide6.QtWidgets import (
+    QWidget, QFrame, QVBoxLayout, QPushButton, QSizePolicy,
+)
+
+
+class _CollapsibleSection(QWidget):
+    """Bold toggle-header + separator + collapsible body."""
+
+    def __init__(self, title: str, parent: QWidget = None) -> None:
+        super().__init__(parent)
+        self.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Maximum)
+        self._title = title
+        self._expanded = True
+
+        root = QVBoxLayout(self)
+        root.setContentsMargins(0, 0, 0, 0)
+        root.setSpacing(0)
+
+        self._toggle_btn = QPushButton(f"▾  {title}")
+        self._toggle_btn.setCheckable(True)
+        self._toggle_btn.setChecked(True)
+        self._toggle_btn.setStyleSheet(
+            "text-align: left; font-weight: bold; padding: 5px 10px;"
+        )
+        self._toggle_btn.clicked.connect(self._on_toggle)
+        root.addWidget(self._toggle_btn)
+
+        sep = QFrame()
+        sep.setFrameShape(QFrame.HLine)
+        sep.setFrameShadow(QFrame.Sunken)
+        root.addWidget(sep)
+
+        self._body = QWidget()
+        self._body_layout = QVBoxLayout(self._body)
+        self._body_layout.setContentsMargins(8, 6, 8, 8)
+        self._body_layout.setSpacing(4)
+        root.addWidget(self._body)
+
+    def body_layout(self) -> QVBoxLayout:
+        return self._body_layout
+
+    def _on_toggle(self, checked: bool) -> None:
+        self._expanded = checked
+        self._body.setVisible(checked)
+        arrow = "▾" if checked else "▸"
+        self._toggle_btn.setText(f"{arrow}  {self._title}")
